Separate models, add Embedded Documents
Separate all the models that were in Estudio in different documents.
Add the Embedded Documents extension.
